Equality now works even if objects travel through the network.
Until now, equality of Objects was equality of their pointers, so if an object
was sent to a service and returned, it would not be equal to the original
object anymore.
An object id is in fact a hash on the pointer passed to the object at the time
of its creation. This hash is a sha1 hash on the concatenation of the machine
uuid, the process uuid and the pointer value. An uuid is used for the process
instead of the process id to avoid different pointers at different times having
the same hash due to process id recycling.
A debug mode also exists where the first number of PtrUid is deterministic
(i.e. it doesn't change at each executions).

Change-Id: I941f67542b9114950bf2e97ef89723caebdeeb18
Reviewed-on: http://gerrit.aldebaran.lan/89024
Reviewed-by: jmonnon <jmonnon@aldebaran.com>
Reviewed-by: jlamotte <jlamotte@aldebaran.com>
Reviewed-by: vincent.palancher <vincent.palancher@external.softbankrobotics.com>
Tested-by: jmonnon <jmonnon@aldebaran.com>