Future: drop onCancel callback after the first call
This is a user-visible change: the callback will only be called
once, while previously, several calls could have occurred.

Yet, this behaviour is not documented, I suspect changing it is ok.

With the new behaviour, arguments captured by the callback are freed
earlier, which is better, and could avoid shared_ptr loops in some
cases.

Change-Id: I2d7b2101ee1d34b6f610564a70838cbd581d461a
Reviewed-on: http://gerrit.aldebaran.lan/84325
Reviewed-by: vincent.palancher <vincent.palancher@external.softbankrobotics.com>
Reviewed-by: jlamotte <jlamotte@aldebaran.com>
Reviewed-by: jmonnon <jmonnon@aldebaran.com>
Tested-by: jlamotte <jlamotte@aldebaran.com>