Adds a log handler routing logs to the debug output in Visual Studio when a debugger is attached.
This is designed to be useful when debugging several processes in parallel.
For example, when launching a parent process spawning several other processes,
Visual Studio can track the child processes and attach the debugger automatically
to them. However in this case only the qiLog output from the parent process is
visible because it is routed only in the console output.
Displaying all the processes logging in the debug output helps tracking what each
process was logging separately and help compare them in one text block.
To make sure this is useful, the launch command line and PID informations
are added to the formatted log messages in this case.

Change-Id: I8f072642b846050de1c3707f3c86c78cfa18c0a4
Reviewed-on: http://gerrit.aldebaran.lan/84717
Reviewed-by: jlamotte <jlamotte@aldebaran.com>
Reviewed-by: vincent.palancher <vincent.palancher@external.softbankrobotics.com>
Tested-by: jlamotte <jlamotte@aldebaran.com>