strand: Executes async callbacks immediately when already in context
When strand is required to execute a function with no delay and the
thread is already in the execution context of the strand, it needs to
execute the function immediately to avoid a deadlock. Until now it
wasn't the case. This patch adds a execNow function in StrandPrivate and
calls it in asyncDelayImpl in the right case.

Change-Id: Ic59f95e5b0830382a16880b6859905acab27e8db
Reviewed-on: http://gerrit.aldebaran.lan/88216
Reviewed-by: vincent.palancher <vincent.palancher@external.softbankrobotics.com>
Tested-by: vincent.palancher <vincent.palancher@external.softbankrobotics.com>
Reviewed-by: jlamotte <jlamotte@aldebaran.com>