gateway: Fixes crashes related to null pointer members
The members _server and _sdClient of the Gateway class are pointers and
as such, the possibility they are null must be handled.

The problem we had was that while the destructor of GatewayPrivate was
calling the destructor of _server (which is a pointer to Session), some
other methods of the class that used _server could be called at the same
time as signal callbacks.

This patch makes sure that we don't use any of these members when they
are null and also reinforces synchronization between method calls.

Change-Id: I9e71a09fbea889850da8ff20d179d24b84d827aa
Reviewed-on: http://gerrit.aldebaran.lan/88218
Reviewed-by: vincent.palancher <vincent.palancher@external.softbankrobotics.com>
Tested-by: vincent.palancher <vincent.palancher@external.softbankrobotics.com>
Reviewed-by: jmonnon <jmonnon@aldebaran.com>
Reviewed-by: jlamotte <jlamotte@aldebaran.com>